Qwen_llama/motia/pyproject.toml
[project]
name = "sales-analytics-motia"
version = "0.1.0"
description = "Motia workflow for Qwen + LLaMA Sales Analytics Assistant"
requires-python = ">=3.10"

dependencies = [
    "motia[otel]==1.0.0rc24",
    "iii-sdk==0.2.0",
    "pydantic>=2.0",
    "requests>=2.32.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]
